// src/app/Models/ContectViewModel.ts
export class ContactViewModel {
    contactId: string;
    entityId: string;
    contactTypeId: number;
    firstName: string;
    lastName: string;
    jobTitle: string;
    emailAddress: string;
    officePhone: string;
    mobilePhone: string;
    statusId: number;
    statusDateTime: Date;
    tenantId: string;
    createdBy: string;
    createDateTime: Date;
    lastUpdatedBy: string;
    lastUpdateDateTime: Date;    
  }